Emitindo JWT pela controller

+++ src/services/identidade/NSE.Identidade.API/Extensions/JwtConfig.cs
@@ -5,6 +5,6 @@
         public string Audience { get; set; }
         public string Issuer { get; set; }
         public string key { get; set; }
-        public int ExpirationTimeHours { get; set; }
+        public double ExpirationTimeHours { get; set; }
     }
 }
